% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00package_help.R
\docType{package}
\name{uscolleges-package}
\alias{uscolleges-package}
\title{uscolleges: Tidy US Scorecard Data}
\description{
A tidy version of the most recent College Scorecards from the US Department
of Education.
}
