% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01data_definitions.R
\docType{data}
\name{uscolleges}
\alias{uscolleges}
\title{Tidy US college scorecard data}
\format{A tibble with 7593 observations of colleges in the United States, and
  622 variables. The variables are described in the
  \code{\link{uscolleges_data_dictionary}}.}
\source{
\url{https://catalog.data.gov/dataset/college-scorecard}
}
\usage{
uscolleges
}
\description{
The most recent US college scorecard from the US Department of Education,
using the dev-friendly column names from the provided data dictionary.
}
\keyword{datasets}
