% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01data_definitions.R
\docType{data}
\name{uscolleges_data_dictionary}
\alias{uscolleges_data_dictionary}
\title{Descriptions of the uscolleges data.}
\format{A tibble with 622 observations of fields, and 4 variables:
  \describe{\item{field}{The field as it appears in uscolleges.}
  \item{API_field}{The field as it appeared in the source data and the
  college scorecard API.} \item{description}{The meaning of the field.}
  \item{category}{The category to which the field belongs. One of root,
  school, admissions, academics, student, cost, aid, completion, and
  repayment.}}}
\source{
\url{https://collegescorecard.ed.gov/data/documentation/}
}
\usage{
uscolleges_data_dictionary
}
\description{
A data dictionary describing the uscolleges data, derived from the dictionary
provided by the US Department of Education.
}
\keyword{datasets}
